+++ spec/unit/sql_retry_spec.rb
@@ -1,0 +1,27 @@
+require 'minitest/autorun'
+require 'minitest/spec'
+require 'active_record'
+require 'lhm/sql_retry'
+require 'trilogy'
+
+describe Lhm::SqlRetry do
+  describe '#retriable_trilogy_errors' do
+    it 'only uses exception classes as retriable keys' do
+      retry_errors = Lhm::SqlRetry.new(nil).send(:retriable_trilogy_errors)
+
+      retry_errors.each_key do |error_class|
+        assert error_class.is_a?(Class), "#{error_class.inspect} is not a class"
+        assert error_class <= Exception, "#{error_class.inspect} is not an exception class"
+      end
+    end
+
+    it 'matches trilogy connection error classes without using the ConnectionError module' do
+      retry_errors = Lhm::SqlRetry.new(nil).send(:retriable_trilogy_errors)
+
+      refute_includes retry_errors.keys, Trilogy::ConnectionError
+      assert_includes retry_errors.keys, Trilogy::BaseConnectionError
+      assert_includes retry_errors.keys, Trilogy::ConnectionClosed
+      assert_includes retry_errors.keys, Trilogy::SSLError
+    end
+  end
+end